Rename 'apiClient' and 'persistentStorage' to 'api-client' and 'persistent-storage'
