+++ browser-extension/src/utils/persistent-storage.ts
